+++ lib/command_builder.dart
@@ -1,10 +1,19 @@
-import 'package:flutter/widgets.dart';
-import 'package:flutter_command/flutter_command.dart';
+part of flutter_command;
 
 class CommandBuilder<TParam, TResult> extends StatelessWidget {
   final Command<TParam, TResult> command;
+
+  /// This builder will be called when the
+  /// command is executed successfully, independent of the return value.
+  final Widget Function(BuildContext context, TParam? param)? onSuccess;
+
+  /// If your command has a return value, you can use this builder to build a widget
+  /// when the command is executed successfully.
   final Widget Function(BuildContext context, TResult data, TParam? param)?
       onData;
+
+  /// If the command has no return value or returns null, this builder will be called when the
+  /// command is executed successfully.
   final Widget Function(BuildContext context, TParam? param)? onNullData;
   final Widget Function(
     BuildContext context,
@@ -20,6 +29,7 @@ class CommandBuilder<TParam, TResult> extends StatelessWidget {
 
   const CommandBuilder({
     required this.command,
+    this.onSuccess,
     this.onData,
     this.onNullData,
     this.whileExecuting,
@@ -29,10 +39,13 @@ class CommandBuilder<TParam, TResult> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    if (command._noReturnValue) {}
     return ValueListenableBuilder<CommandResult<TParam?, TResult>>(
       valueListenable: command.results,
       builder: (context, result, _) {
         return result.toWidget(
+          onSuccess: (paramData) =>
+              onSuccess?.call(context, paramData) ?? const SizedBox(),
           onData: (data, paramData) =>
               onData?.call(context, data, paramData) ?? const SizedBox(),
           onNullData: (paramData) =>
@@ -59,19 +72,25 @@ class CommandBuilder<TParam, TResult> extends StatelessWidget {
 extension ToWidgeCommandResult<TParam, TResult>
     on CommandResult<TParam, TResult> {
   Widget toWidget({
-    required Widget Function(TResult result, TParam? param) onData,
+    Widget Function(TResult result, TParam? param)? onData,
+    Widget Function(TParam? param)? onSuccess,
     Widget Function(TParam? param)? onNullData,
     Widget Function(TResult? lastResult, TParam? param)? whileExecuting,
     Widget Function(Object error, TResult? lastResult, TParam? param)? onError,
   }) {
+    assert(onData != null || onSuccess != null,
+        'You have to provide at least a builder for onData or onSuccess');
     if (error != null) {
       return onError?.call(error!, data, paramData) ?? const SizedBox();
     }
     if (isExecuting) {
       return whileExecuting?.call(data, paramData) ?? const SizedBox();
     }
+    if (onSuccess != null) {
+      return onSuccess.call(paramData);
+    }
     if (data != null) {
-      return onData(data as TResult, paramData);
+      return onData?.call(data as TResult, paramData) ?? const SizedBox();
     } else {
       return onNullData?.call(paramData) ?? const SizedBox();
     }